+++ activerecord/lib/arel/nodes/delete_statement.rb
@@ -3,7 +3,7 @@
 module Arel # :nodoc: all
   module Nodes
     class DeleteStatement < Arel::Nodes::Node
-      attr_accessor :relation, :wheres, :groups, :havings, :orders, :limit, :offset, :comment, :key
+      attr_accessor :relation, :wheres, :groups, :havings, :orders, :limit, :offset, :comment, :key, :for_portion_of
 
       def initialize(relation = nil, wheres = [])
         super()
@@ -16,6 +16,7 @@ module Arel # :nodoc: all
         @offset = nil
         @comment = nil
         @key = nil
+        @for_portion_of = nil
       end
 
       def initialize_copy(other)
@@ -25,7 +26,7 @@ module Arel # :nodoc: all
       end
 
       def hash
-        [self.class, @relation, @wheres, @orders, @limit, @offset, @comment, @key].hash
+        [self.class, @relation, @wheres, @orders, @limit, @offset, @comment, @key, @for_portion_of].hash
       end
 
       def eql?(other)
@@ -38,7 +39,8 @@ module Arel # :nodoc: all
           self.limit == other.limit &&
           self.offset == other.offset &&
           self.comment == other.comment &&
-          self.key == other.key
+          self.key == other.key &&
+          self.for_portion_of == other.for_portion_of
       end
       alias :== :eql?
     end